pyproject.toml
[tool.poetry]
name = "Python Micro Framework Data FastAPI Example"
version = "0.1.0"
description = ""
authors = ["Neon K.I.D <contact@neonkid.xyz>"]

[tool.poetry.dependencies]
python = "^3.8.10"
fastapi = "^0.70.1"
uvicorn = {extras = ["standard"], version = "^0.16.0"}
dependency-injector = "^4.37.0"
asyncpg = "^0.25.0"
loguru = "^0.5.3"
orjson = "^3.6.5"
python-mf-data = {extras = ["rdb"], version = "^0.2.4"}

[tool.poetry.dev-dependencies]
pytest-asyncio = "^0.16.0"
mypy = "^0.930"
httpx = "^0.21.1"
asgi-lifespan = "^1.0.1"
Faker = "^9.2.0"
psycopg2-binary = "^2.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
